Moduralized Roster.js and switched it with a renderRoster function that adds functionality for team/player adding page by calling modules.  Called that in main to test
